Move set of built-in activities and store utility to stores
